backend/requirements.txt
fastapi
uvicorn
python-dotenv
ffmpeg-python
openai==1.93.0
chromadb==1.0.15
langchain==0.3.26
langchain-openai==0.3.27
sentence-transformers==5.0.0
langchain_community
langchain_chroma
